recipe-api/requirements.txt
Django==2.2.12
djangorestframework==3.11.0
flake8==3.7.9
parameterized==0.7.4
